* [CHANGE]   * fix the Carp::xxx() calls Test::Smoke::Patcher   + add support for patch descriptions     + run $^X -x patchlevel.h "$descr"   + add tests for the descriptions in patchlevel.h   * update the snapshots with newer patchlevel.h to test the descriptions
git-svn-id: https://source.test-smoke.org/svn/Test-Smoke@883 f9e6db9c-56bf-0310-a82a-abbfb51efdf0

+++ t/patcher.t
@@ -11,7 +11,7 @@ use lib $findbin;
 use TestLib;
 use Cwd;
 
-use Test::More tests => 32;
+use Test::More tests => 35;
 BEGIN { use_ok( 'Test::Smoke::Patcher' ) };
 my $verbose = exists $ENV{SMOKE_VERBOSE} ? $ENV{SMOKE_VERBOSE} : 0;
 
@@ -100,7 +100,7 @@ SKIP: { # test Test::Smoke::Patcher->patch_single()
 }
 
 SKIP: { # Test multi mode
-    my $to_skip = 9;
+    my $to_skip = 12;
 
     skip "No patch program or test-patch found", $to_skip
         unless $patch && -e $testpatch;
@@ -153,6 +153,15 @@ EOPINFO
     $newfile = get_file(qw( t perl patchme.txt ));
     unlike( $newfile, '/^VERSION == 20001$/m', "Conent OK" );
     1 while unlink $pinfo;
+
+    my $descr = '[PATCH] just testing comments';
+    eval { $patcher->patch_single( $relpatch, '', $descr ) };
+    ok ! $@, "Patch applied($descr) $@";
+    $newfile = get_file(qw( t perl patchme.txt ));
+    like( $newfile, '/^VERSION == 20001$/m', "Conent OK" );
+    my $plevel = get_file(qw( t perl patchlevel.h ));
+    like $plevel, qq{m/^\\s*,"\Q$descr\E"/m},
+         "Description added to patchlevel.h";
 }
 
 {
@@ -167,6 +176,8 @@ EOPINFO
 }
 
 END {
-    rmtree( File::Spec->catdir(qw( t perl )) );
-    1 while unlink $testpatch;
+    unless ( $ENV{SMOKE_DEBUG} ) {
+        rmtree( File::Spec->catdir(qw( t perl )) );
+        1 while unlink $testpatch;
+    }
 }
